+++ radix-engine/src/engine/mod.rs
@@ -4,8 +4,8 @@ mod call_frame;
 mod errors;
 mod execution_trace;
 mod kernel;
+mod logger_module;
 mod module;
-mod module_logger;
 mod native_interpreter;
 mod node;
 mod node_properties;
@@ -21,8 +21,8 @@ pub use call_frame::CallFrame;
 pub use errors::*;
 pub use execution_trace::*;
 pub use kernel::*;
+pub use logger_module::*;
 pub use module::*;
-pub use module_logger::*;
 pub use native_interpreter::NativeInterpreter;
 pub use node::*;
 pub use node_properties::*;
